+++ bittytax/conv/parsers/electrum.py
@@ -4,7 +4,7 @@
 from decimal import Decimal
 
 from ...config import config
-from ...record import TransactionRecordBase as TransactionOutRecord
+from ..out_record import TransactionOutRecord
 from ..dataparser import DataParser
 from ..exceptions import UnknownCryptoassetError